Add types key to package.json
On the latest Typescript, I couldn't get the Typescript compiler or VS Code to recognize that the `ts-pptx` library had a typings file. 

Adding this `types` key [seems to be required](https://www.typescriptlang.org/docs/handbook/declaration-files/publishing.html) if the `*.d.ts` file isn't at the root of the project.

Thanks for maintaining this project!